Change plugin configurations to be provider specific
Plugin configuration view depends now on which provider is used. It
is therefore more logical for the configurations to be near the
providers.

Added a Configuration interface which desribes the template and
the fields used on the template. This way they both are close together
and they can be configured separately for each plugin type.

Currently there is a generic default implementation and a Gerrit
implementation of the Configuration interface.